# openapi-v1.yaml
openapi: 3.1.0
info:
  title: ARB AI Agent API
  description: Architecture Review Board AI Agent Backend
  version: 1.0.0
paths:
...
  /api/v1/submissions:
    get:
      tags:
      - v1 / submissions
      summary: Get Submissions
      description: Get all ARB submissions
      operationId: get_submissions_api_v1_submissions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - v1 / submissions
      summary: Create Submission
      description: Create a new ARB submission
      operationId: create_submission_api_v1_submissions_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ARBSubmission'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/submissions/{submission_id}:
    get:
      tags:
      - v1 / submissions
      summary: Get Submission
      description: Get a specific ARB submission
      operationId: get_submission_api_v1_submissions__submission_id__get
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - v1 / submissions
      summary: Update Submission
      description: Update an existing ARB submission
      operationId: update_submission_api_v1_submissions__submission_id__put
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ARBSubmission'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/submissions/{submission_id}/submit:
    post:
      tags:
      - v1 / submissions
      summary: Submit Submission
      description: Submit ARB submission for review
      operationId: submit_submission_api_v1_submissions__submission_id__submit_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/submissions/{submission_id}/artefacts:
    post:
      tags:
      - v1 / submissions
      summary: Upload Artefact
      description: Upload an artefact for a submission
      operationId: upload_artefact_api_v1_submissions__submission_id__artefacts_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_artefact_api_v1_submissions__submission_id__artefacts_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/submissions/{submission_id}/integration-catalogue:
    post:
      tags:
      - v1 / submissions
      summary: Upload Integration Catalogue
      description: Upload integration catalogue (Excel/CSV)
      operationId: upload_integration_catalogue_api_v1_submissions__submission_id__integration_catalogue_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_integration_catalogue_api_v1_submissions__submission_id__integration_catalogue_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/v2/submissions:
    get:
      tags:
      - v2 / submissions
      summary: Get Submissions
      description: Get all ARB submissions
      operationId: get_submissions_api_v2_submissions_get
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - v2 / submissions
      summary: Create Submission
      description: Create a new ARB submission
      operationId: create_submission_api_v2_submissions_post
      parameters:
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ARBSubmission'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/submissions/{submission_id}:
    get:
      tags:
      - v2 / submissions
      summary: Get Submission
      description: Get a specific ARB submission
      operationId: get_submission_api_v2_submissions__submission_id__get
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    put:
      tags:
      - v2 / submissions
      summary: Update Submission
      description: Update an existing ARB submission
      operationId: update_submission_api_v2_submissions__submission_id__put
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ARBSubmission'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/submissions/{submission_id}/submit:
    post:
      tags:
      - v2 / submissions
      summary: Submit Submission
      description: Submit ARB submission for review
      operationId: submit_submission_api_v2_submissions__submission_id__submit_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/submissions/{submission_id}/artefacts:
    post:
      tags:
      - v2 / submissions
      summary: Upload Artefact
      description: Upload an artefact for a submission
      operationId: upload_artefact_api_v2_submissions__submission_id__artefacts_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_artefact_api_v2_submissions__submission_id__artefacts_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v2/submissions/{submission_id}/integration-catalogue:
    post:
      tags:
      - v2 / submissions
      summary: Upload Integration Catalogue
      description: Upload integration catalogue (Excel/CSV)
      operationId: upload_integration_catalogue_api_v2_submissions__submission_id__integration_catalogue_post
      parameters:
      - name: submission_id
        in: path
        required: true
        schema:
          type: string
          title: Submission Id
      - name: authorization
        in: header
        required: false
        schema:
          anyOf:
          - type: string
          - type: 'null'
          title: Authorization
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Body_upload_integration_catalogue_api_v2_submissions__submission_id__integration_catalogue_post'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
    ARBSubmission:
      properties:
        id:
          anyOf:
          - type: string
          - type: 'null'
          title: Id
        project_name:
          type: string
          title: Project Name
        solution_architect_id:
          type: string
          title: Solution Architect Id
        status:
          type: string
          title: Status
          default: draft
        created_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Created Date
        submitted_date:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Submitted Date
        problem_statement:
          type: string
          title: Problem Statement
        stakeholders:
          items:
            type: string
          type: array
          title: Stakeholders
        business_drivers:
          items:
            type: string
          type: array
          title: Business Drivers
        growth_plans:
          anyOf:
          - type: string
          - type: 'null'
          title: Growth Plans
        application_architecture:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        integration_architecture:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        data_architecture:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        security_architecture:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        infrastructure_architecture:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        devsecops:
          anyOf:
          - $ref: '#/components/schemas/DomainSection'
          - type: 'null'
        nfr_criteria:
          items:
            $ref: '#/components/schemas/NFRCriteria'
          type: array
          title: Nfr Criteria
        overall_progress:
          type: number
          title: Overall Progress
          default: 0.0
      type: object
      required:
      - project_name
      - solution_architect_id
      - problem_statement
      - stakeholders
      - business_drivers
      - nfr_criteria
      title: ARBSubmission
    Artefact:
      properties:
        id:
          type: string
          title: Id
        file_name:
          type: string
          title: File Name
        file_type:
          type: string
          title: File Type
        file_size:
          type: integer
          title: File Size
        system_label:
          type: string
          title: System Label
        upload_date:
          type: string
          format: date-time
          title: Upload Date
        file_path:
          type: string
          title: File Path
      type: object
      required:
      - id
      - file_name
      - file_type
      - file_size
      - system_label
      - upload_date
      - file_path
      title: Artefact
...
    Body_upload_artefact_api_v1_submissions__submission_id__artefacts_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_artefact_api_v1_submissions__submission_id__artefacts_post
...
    Body_upload_artefact_api_v2_submissions__submission_id__artefacts_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_artefact_api_v2_submissions__submission_id__artefacts_post
    Body_upload_integration_catalogue_api_v1_submissions__submission_id__integration_catalogue_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_integration_catalogue_api_v1_submissions__submission_id__integration_catalogue_post
    Body_upload_integration_catalogue_api_v2_submissions__submission_id__integration_catalogue_post:
      properties:
        file:
          type: string
          format: binary
          title: File
      type: object
      required:
      - file
      title: Body_upload_integration_catalogue_api_v2_submissions__submission_id__integration_catalogue_post
    ChecklistItem:
      properties:
        id:
          type: string
          title: Id
        question:
          type: string
          title: Question
        answer:
          $ref: '#/components/schemas/ChecklistOption'
        evidence_notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence Notes
      type: object
      required:
      - id
      - question
      - answer
      title: ChecklistItem
    ChecklistOption:
      type: string
      enum:
      - compliant
      - non_compliant
      - partial
      - na
      title: ChecklistOption
...
    DomainSection:
      properties:
        domain:
          type: string
          title: Domain
        checklist_items:
          items:
            $ref: '#/components/schemas/ChecklistItem'
          type: array
          title: Checklist Items
        integration_catalogue:
          anyOf:
          - items:
              $ref: '#/components/schemas/IntegrationCatalogueItem'
            type: array
          - type: 'null'
          title: Integration Catalogue
        artefacts:
          items:
            $ref: '#/components/schemas/Artefact'
          type: array
          title: Artefacts
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - domain
      - checklist_items
      - artefacts
      title: DomainSection
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
    IntegrationCatalogueItem:
      properties:
        id:
          type: string
          title: Id
        system_name:
          type: string
          title: System Name
        interface_type:
          type: string
          title: Interface Type
        protocol:
          type: string
          title: Protocol
        data_format:
          type: string
          title: Data Format
        frequency:
          type: string
          title: Frequency
        security:
          type: string
          title: Security
        notes:
          anyOf:
          - type: string
          - type: 'null'
          title: Notes
      type: object
      required:
      - id
      - system_name
      - interface_type
      - protocol
      - data_format
      - frequency
      - security
      title: IntegrationCatalogueItem
...
    NFRCriteria:
      properties:
        category:
          type: string
          title: Category
        criteria:
          type: string
          title: Criteria
        target_value:
          type: string
          title: Target Value
        actual_value:
          type: string
          title: Actual Value
        score:
          type: integer
          maximum: 5.0
          minimum: 1.0
          title: Score
        evidence:
          anyOf:
          - type: string
          - type: 'null'
          title: Evidence
      type: object
      required:
      - category
      - criteria
      - target_value
      - actual_value
      - score
      title: NFRCriteria
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError